+++ middleware/arduino.js
@@ -1,9 +1,7 @@
-const arduinoModel = require('../modules/arduino');
-
-
 exports.createArduino = (req, res) => {
     const data = req.body;
-    arduinoModel.createArduino(data, (err, result) => {
+    const sql = 'INSERT INTO arduino SET ?';
+    dbPool.query(sql, data, (err, result) => {
         if (err) {
             console.error('Error creating Arduino device:', err);
             return res.status(500).send('Error creating Arduino device');
@@ -13,7 +11,8 @@ exports.createArduino = (req, res) => {
 };
 
 exports.getAllArduinos = (req, res) => {
-    arduinoModel.getAllArduinos((err, results) => {
+    const sql = 'SELECT * FROM arduino';
+    dbPool.query(sql, (err, results) => {
         if (err) {
             console.error('Error retrieving Arduino devices:', err);
             return res.status(500).send('Error retrieving Arduino devices');
@@ -23,7 +22,8 @@ exports.getAllArduinos = (req, res) => {
 };
 
 exports.getArduinoById = (req, res) => {
-    arduinoModel.getArduinoById(req.params.id, (err, result) => {
+    const sql = 'SELECT * FROM arduino WHERE id = ?';
+    dbPool.query(sql, [req.params.id], (err, result) => {
         if (err) {
             console.error('Error retrieving Arduino device:', err);
             return res.status(500).send('Error retrieving Arduino device');
@@ -37,7 +37,8 @@ exports.getArduinoById = (req, res) => {
 
 exports.updateArduino = (req, res) => {
     const data = req.body;
-    arduinoModel.updateArduino(req.params.id, data, (err, result) => {
+    const sql = 'UPDATE arduino SET ? WHERE id = ?';
+    dbPool.query(sql, [data, req.params.id], (err, result) => {
         if (err) {
             console.error('Error updating Arduino device:', err);
             return res.status(500).send('Error updating Arduino device');
@@ -50,7 +51,8 @@ exports.updateArduino = (req, res) => {
 };
 
 exports.deleteArduino = (req, res) => {
-    arduinoModel.deleteArduino(req.params.id, (err, result) => {
+    const sql = 'DELETE FROM arduino WHERE id = ?';
+    dbPool.query(sql, [req.params.id], (err, result) => {
         if (err) {
             console.error('Error deleting Arduino device:', err);
             return res.status(500).send('Error deleting Arduino device');
