+++ middleware/environmental_avg.js
@@ -1,62 +1,86 @@
-const environmentalAvgModel = require('../modules/environmental_avg');
 
-exports.createEnvironmentalAvgData = (req, res) => {
-    const data = req.body;
-    environmentalAvgModel.createEnvironmentalAvgData(data, (err, result) => {
-        if (err) {
-            console.error('Error creating environmental average data:', err);
-            return res.status(500).send('Error creating environmental average data');
-        }
-        res.status(201).send(result);
+const getAllEnvironmentalAvgData = (callback) => {
+    dbPool.query('SELECT * FROM environmental_data_avg', (err, rows) => {
+        if (err) return callback(err);
+        callback(null, rows);
     });
 };
 
-exports.getAllEnvironmentalAvgData = (req, res) => {
-    environmentalAvgModel.getAllEnvironmentalAvgData((err, results) => {
-        if (err) {
-            console.error('Error retrieving environmental average data:', err);
-            return res.status(500).send('Error retrieving environmental average data');
-        }
-        res.send(results);
+const getEnvironmentalAvgDataById = (id, callback) => {
+    dbPool.query('SELECT * FROM environmental_data_avg WHERE id = ?', [id], (err, rows) => {
+        if (err) return callback(err);
+        callback(null, rows[0]);
     });
 };
 
-exports.getEnvironmentalAvgDataById = (req, res) => {
-    environmentalAvgModel.getEnvironmentalAvgDataById(req.params.id, (err, result) => {
-        if (err) {
-            console.error('Error retrieving environmental average data by ID:', err);
-            return res.status(500).send('Error retrieving environmental average data by ID');
-        }
-        if (result.length === 0) {
-            return res.status(404).send('Environmental average data not found');
-        }
-        res.send(result);
-    });
+const createEnvironmentalAvgData = (data, callback) => {
+    const {
+        device_id, uv_radiation, uv_radiation_max, uv_radiation_min,
+        light, light_max, light_min, air_temperature, air_temperature_max,
+        air_temperature_min, air_humidity, air_humidity_max, air_humidity_min,
+        soil_humidity, soil_humidity_max, soil_humidity_min, plant_ID
+    } = data;
+    const measurement_date = new Date(); // זמן עכשיו
+
+
+    const avg = [
+        device_id, uv_radiation, uv_radiation_max, uv_radiation_min,
+        light, light_max, light_min, air_temperature, air_temperature_max,
+        air_temperature_min, air_humidity, air_humidity_max, air_humidity_min,
+        soil_humidity, soil_humidity_max, soil_humidity_min, plant_ID, measurement_date
+    ];
+
+    const query = `
+        INSERT INTO environmental_data_avg (
+            device_id, uv_radiation, uv_radiation_max, uv_radiation_min,
+            light, light_max, light_min, air_temperature, air_temperature_max,
+            air_temperature_min, air_humidity, air_humidity_max, air_humidity_min,
+            soil_humidity, soil_humidity_max, soil_humidity_min, plant_ID,
+            measurement_date
+        ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+    `;
+
+    dbPool.query(query, avg, callback);
 };
 
-exports.updateEnvironmentalAvgData = (req, res) => {
-    const data = req.body;
-    environmentalAvgModel.updateEnvironmentalAvgData(req.params.id, data, (err, result) => {
-        if (err) {
-            console.error('Error updating environmental average data:', err);
-            return res.status(500).send('Error updating environmental average data');
-        }
-        if (result.affectedRows === 0) {
-            return res.status(404).send('Environmental average data not found');
-        }
-        res.send(result);
-    });
+const updateEnvironmentalAvgData = (id, data, callback) => {
+    const {
+        device_id, uv_radiation, uv_radiation_max, uv_radiation_min,
+        light, light_max, light_min, air_temperature, air_temperature_max,
+        air_temperature_min, air_humidity, air_humidity_max, air_humidity_min,
+        soil_humidity, soil_humidity_max, soil_humidity_min, plant_ID, measurement_date
+    } = data;
+
+
+    const avg = [
+        device_id, uv_radiation, uv_radiation_max, uv_radiation_min,
+        light, light_max, light_min, air_temperature, air_temperature_max,
+        air_temperature_min, air_humidity, air_humidity_max, air_humidity_min,
+        soil_humidity, soil_humidity_max, soil_humidity_min, plant_ID, measurement_date, id
+    ];
+
+    const query = `
+        UPDATE environmental_data_avg SET
+            device_id = ?, uv_radiation = ?, uv_radiation_max = ?, uv_radiation_min = ?,
+            light = ?, light_max = ?, light_min = ?, air_temperature = ?, air_temperature_max = ?,
+            air_temperature_min = ?, air_humidity = ?, air_humidity_max = ?, air_humidity_min = ?,
+            soil_humidity = ?, soil_humidity_max = ?, soil_humidity_min = ?, plant_ID = ?,
+            measurement_date = ?
+        WHERE id = ?
+    `;
+
+    dbPool.query(query, avg, callback);
 };
 
-exports.deleteEnvironmentalAvgData = (req, res) => {
-    environmentalAvgModel.deleteEnvironmentalAvgData(req.params.id, (err, result) => {
-        if (err) {
-            console.error('Error deleting environmental average data:', err);
-            return res.status(500).send('Error deleting environmental average data');
-        }
-        if (result.affectedRows === 0) {
-            return res.status(404).send('Environmental average data not found');
-        }
-        res.send(result);
-    });
+const deleteEnvironmentalAvgData = (id, callback) => {
+    const sql = 'DELETE FROM environmental_data_avg WHERE id = ?';
+    dbPool.query(sql, [id], callback);
+};
+
+module.exports = {
+    getAllEnvironmentalAvgData,
+    getEnvironmentalAvgDataById,
+    createEnvironmentalAvgData,
+    updateEnvironmentalAvgData,
+    deleteEnvironmentalAvgData
 };
